branch:	HACKT-00-01-04-main-00 (main trunk) rev:	HACKT-00-01-04-main-00-80-57-merged-channel-01-09
  Merged channel-01(-09) branch.
  Summary: channel aliases now have directional connection
	semantics -- each alias (set) having producer/consumer flags.
	As connections are made, flag are checked and updated
	for conflicts.  Final checking reports dangling channels.
	Also includes (disabled) patch for propagating connection
	summary information from formals to actuals, including
	channel direction and relaxed actuals.  Enabling this patch
	will obsolete the unroll phase.
  Minor fixes: unresolved data-type of pint is int<0> (magic)

+++ test/update-createstderr.sh
@@ -23,7 +23,16 @@ done
 
 for i
 do
-	cp -f $i.createdump-filter $srcdir/$i.createstderr
-	echo "Updated $srcdir/$i.createstderr"
+	if test -f $i.createdump-filter
+	then
+		cp -f $i.createdump-filter $srcdir/$i.createstderr
+		echo "Updated $srcdir/$i.createstderr"
+	elif test -f $i.createfaildump
+	then
+		cp -f $i.createfaildump $srcdir/$i.createstderr
+		echo "Updated $srcdir/$i.createstderr"
+	else
+		echo "$i.createdump-filter or $i.createfaildump required but missing."
+	fi
 done
